requirements.txt
streamlit
pandas
requests
python-dotenv
langchain
langchain-google-genai
